Send Fax Remove message text as not working
Removed the message text in the Send Fax screen as this did not work, it
has only been marked as hidden so it does not show this to the end user.
Also made some UI changes on received faxes - delete and view buttons
are swapped over so the Delete button is always first and shows on the
left hand side for continuity purposes.